+++ CHANGELOG.md
@@ -11,12 +11,17 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+- Bump stackable-operator to 0.108.0, strum to 0.28, snafu to 0.9, and ldap3 to 0.12 ([#685], [#688]).
 - Gracefully shutdown all concurrent tasks by forwarding the SIGTERM signal ([#674]).
 - OLM deployer doesn't add owner references to cluster scoped objects anymore ([#667]).
   Owner references ensure that objects are garbage collected by OpenShift upon operator removal but they cause problems when the operator is updated.
   This means that cluster wide objects are not removed anymore when the operator is uninstalled.
   This behaviour is in line with the default behaviour of Helm and OLM.
 - Bump testing-tools to `0.3.0-stackable0.0.0-dev` ([#671]).
+
+### Fixed
+
+- Fix "404 page not found" error for the initial object list ([#687], [#688]).
 
 ### Removed
 
@@ -44,6 +49,9 @@ All notable changes to this project will be documented in this file.
 [#671]: https://github.com/stackabletech/secret-operator/pull/671
 [#674]: https://github.com/stackabletech/secret-operator/pull/674
 [#679]: https://github.com/stackabletech/secret-operator/pull/679
+[#685]: https://github.com/stackabletech/secret-operator/pull/685
+[#687]: https://github.com/stackabletech/secret-operator/pull/687
+[#688]: https://github.com/stackabletech/secret-operator/pull/688
 
 ## [25.11.0] - 2025-11-07